Set user back to pending if initial payment fails. #1060

+++ includes/gateways/class-rcp-payment-gateway-paypal-express.php
@@ -518,13 +518,19 @@ class RCP_Payment_Gateway_PayPal_Express extends RCP_Payment_Gateway {
 
 				if( ! $member->just_upgraded() ) {
 
-					// user is marked as cancelled but retains access until end of term
-					$member->cancel();
+					if( isset( $posted['initial_payment_status'] ) && 'Failed' == $posted['initial_payment_status'] ) {
+						// Initial payment failed, so set the user back to pending.
+						$member->set_status( 'pending' );
+						$member->add_note( __( 'Initial payment failed in PayPal Express.', 'rcp' ) );
+					} else {
+						// user is marked as cancelled but retains access until end of term
+						$member->cancel();
 
-					// set the use to no longer be recurring
-					delete_user_meta( $user_id, 'rcp_paypal_subscriber' );
+						// set the use to no longer be recurring
+						delete_user_meta( $user_id, 'rcp_paypal_subscriber' );
 
-					do_action( 'rcp_ipn_subscr_cancel', $user_id );
+						do_action( 'rcp_ipn_subscr_cancel', $user_id );
+					}
 
 					die( 'successful recurring_payment_profile_cancel' );
